simplify nlu,dst,policy,nlg interface
