+++ tatk/dst/__init__.py
@@ -1,1 +1,1 @@
-from tatk.dst.state_tracker import DST
+from tatk.dst.dst import DST
