+++ tatk/nlg/nlg.py
@@ -1,5 +1,4 @@
 """Natural Language Generation Interface"""
-from abc import abstractmethod
 from tatk.util.module import Module
 
 
@@ -16,18 +15,4 @@ class NLG(Module):
             utterance (str):
                 A natural langauge utterance.
         """
-        return self.generate_batch([action])[0]
-
-    # @abstractmethod
-    def generate_batch(self, batch_action):
-        """Generate natural language utterances conditioned on a batch of dialog acts.
-
-        Args:
-            batch_action (list of list of tuples):
-                A batch of dialog acts
-
-        Returns:
-            batch_utterance (list of string):
-                Natural language utterances.
-        """
-        return [[]]
+        return ''
