+++ tatk/policy/policy.py
@@ -1,5 +1,4 @@
 """Policy Interface"""
-from abc import abstractmethod
 from tatk.util.module import Module
 
 
@@ -17,18 +16,4 @@ class Policy(Module):
             action (list of tuples):
                 The next dialog action.
         """
-        return self.predict_batch([state])[0]
-
-    # @abstractmethod
-    def predict_batch(self, batch_state):
-        """Predict actions given a batch of dialog states.
-
-        Args:
-            batch_state (list of tuple or dict):
-                when the DST and Policy module are separated, the type of state is tuple.
-                else when they are aggregated together, the type of state is dict (dialog act).
-        Returns:
-            batch_action (list of list of tuples):
-                The next dialog action.
-        """
-        pass
+        return []
